Complete rewrite of game finding logic
The previous logic (a carryover from the project this was forked from)
was broken in multiple ways. It didn't account for non-root Steam
libraries, nor did it account for filesystem variations such as missing
symlinks.

The new logic corrects all of this by simply using GNU `find` to search
for Steam appmanifest files by APPID.